backend/bckend/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-wsgi==1.0b1
djangorestframework==3.14.0
drf-yasg==1.21.7
importlib-metadata==7.0.1
inflection==0.5.1
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.10.0
sqlparse==0.4.4
uritemplate==4.1.1
WebOb==1.8.7
zipp==3.17.0
